requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
SQLAlchemy==2.0.20
PyMySQL==1.0.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.3.0
python-dotenv